refactor(renderer/api/announcement): move interfaces to co-located types/ folder
